Revise API URLs, change PUT /auth/login to POST, revise tests
