+++ backend/app/routers/auth.py
@@ -1,63 +1,71 @@
-"""Route handler to authenticate a User."""
+"""Router endpoints for account management, including change password."""
 
 import logging
-from pathlib import Path
 from typing import Annotated
 from fastapi import APIRouter, Depends, status, HTTPException, Request, Response
 from fastapi.security.oauth2 import OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
 
-from app.core import database, security
+from app.core import database
 from app.data_access import user_dao
+from app.routers.base import API_PREFIX
+from app.routers.login import validate_login
 from app.utils import jwt, oauth2
-from app import schemas
-from fastapi.responses import HTMLResponse
+from app import models, schemas
 
-router = APIRouter(tags=['Authentication'])
-
-
-@router.get('/login', response_class=HTMLResponse)
-async def loginform():
-    """Return a login form in html."""
-    file_path = Path("app/forms/login.html")
-    if not file_path.exists():
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Login form {str(file_path.absolute())} not found"
-                            )
-    return file_path.read_text(encoding="utf-8")
+router = APIRouter(prefix=API_PREFIX, tags=['Authentication API for WS'])
 
 
-@router.post('/login', response_model=schemas.Token)
-async def login(form_data: Annotated[OAuth2PasswordRequestForm, Depends()],
-                session: Session = Depends(database.db.get_session)):
-    """Authenticate user and return a JWT token for this session.
+@router.post('/auth/password', status_code=status.HTTP_204_NO_CONTENT)
+async def change_password(
+        form_data: Annotated[OAuth2PasswordRequestForm, Depends()],
+        session: Session = Depends(database.db.get_session),
+        current_user: models.User = Depends(oauth2.get_current_user)
+        ):
+    """Change password for the currently authenticated user,\
+       where the password is submitted as form-data in a field named 'password'.
 
-    :param form_data: data from login form, containing 'username' and 'password' fields
-
-    `form_data` also has fields for `grant_type`, (optional) `scope`, `client_id`, and `client_secret`
-    The `grant_type` should always be `password` (the OAuth2 flow).
+       The 'username' used in form must be the email of the currently authenticated user.
     """
-    # username is really the email address. In our models, 'username' is arbitrary and not unique.
     try:
-        email = form_data.username
-        password = form_data.password
-    except Exception:
-        logging.warning(f"Login failed for {email}. POST form data invalid.")
-        # Response should probably be 422 Unprocessable Entity
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
-                            detail="Missing user email or password"
-                            )
-    access_token = await validate_login(email, password, session)
-    return {"access_token": access_token, "token_type": "bearer"}
+        # validate password from form_data
+        schemas.PasswordCreate(password=form_data.password)
+    except ValueError as ex:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(ex))
+    user_password = await user_dao.set_password(session, current_user, form_data.password)  # noqa: F841
+    return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
-@router.put('/login', response_model=schemas.Token)
+@router.put('/auth/password', status_code=status.HTTP_204_NO_CONTENT)
+async def update_password(
+        password_request: schemas.PasswordStr,
+        session: Session = Depends(database.db.get_session),
+        current_user: models.User = Depends(oauth2.get_current_user)
+        ):
+    """Change password for the authenticated user, for request submitted as JSON data.
+
+    Returns:
+    - 204 No Content on success (recommended best practice)
+    - 400 Bad Request if request does not contain required data
+    - 401 Unauthorized if token invalid
+    - 422 Validation Errors if the password does not satisfy validation rules,
+    such as at least 1 uppercase letter, 1 lowercase letter, and 1 numeric char
+    """
+    # Instead of returning status 204, it's acceptable to return 200 OK with a text body such as:
+    # { "status": "success", "message": "Password updated" }
+    # but do not return the hashed password.
+    new_password = password_request.get_secret_value()
+    user_password = await user_dao.set_password(session, current_user, new_password)  # noqa: F841
+    return Response(status_code=status.HTTP_204_NO_CONTENT)
+
+
+@router.post('/auth/login', response_model=schemas.Token)
 async def login_json(login_data: schemas.LoginData,
                      request: Request,
                      session: Session = Depends(database.db.get_session)):
     """Authenticate user using JSON input values and return a JWT token.
 
-    :param request: PUT request containing username=value, password=value
+    :param request: POST request containing username=value, password=value
     :returns: JSON access token. Format of body is:
               {"access_token": "jwt-token-value", "token_type": "bearer" }
 
@@ -118,41 +126,3 @@ async def validate_token(token: str = Depends(oauth2.oauth2_scheme)):
         raise HTTPException(status_code=401, detail="Token expired")
     except jwt.JWTError:
         raise HTTPException(status_code=401, detail="Invalid token")
-
-
-async def validate_login(email: str, password: str, session: Session) -> str:
-    """Validate user credentials and return a JWT access token.
-
-    If any errors, raises HTTPException.
-    """
-    if not email or not password:
-        logging.warning(f"Login failed for {email}. Missing username or password.")
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
-                      detail="Username and password may not be empty.")
-    user = await user_dao.get_by_email(session, email=email)
-    if not user:
-        logging.warning(f"Login failed for {email}. Unknown user.")
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid Credentials",
-            headers={"WWW-Authenticate": "Bearer"}
-            )
-    hashed_password = await user_dao.get_password(session, user)
-    if not hashed_password:
-        logging.warning(f"Login failed for {email}. User has no local password.")
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="User does not have local password credential",
-            headers={"WWW-Authenticate": "Bearer"}
-            )
-    if not security.verify_password(hashed_password=hashed_password, plain_password=password):
-        logging.warning(f"Login failed for {email} with {password}. Invalid credentials.")
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid Credentials",
-            headers={"WWW-Authenticate": "Bearer"}
-            )
-    # create and return a token
-    access_token = jwt.create_access_token(data={"user_id": user.id})
-    logging.info(f"Login success for {email} Access token granted.")
-    return access_token
